requirements.txt
flask
reportlab
PyMuPDF
Pillow
